refactor: move funct to utils
